Warn when unused ExitWhenReady config option is set
This option is exposed as a public field in the sidecar config, so
it's not safe to remove it in case someone might be extending this
code. So update the comments to indicate that it is ignored, and
emit a warning if it is found to be set to true.

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>