Warnings about ignored settings when daemon_mode=false
Presently spiffe-helper ignores the settings cmd, pid_file_name
and renew_signal silently if daemon_mode=false.

This could lead to misconfiguration where the helper is not doing what
the user expects. It also makes it more difficult to give these options
a meaning in non-daemon (one-shot) mode in future, if existing user
configurations might contain invalid uses of them that presently have no
effect.

Raise an error if pid_file_name is configured with daemon_mode=false.
This option is new and isn't documented yet so the risk of people having
configurations that set it and daemon_mode=false is minimal.

For cmd and renew_signal, log a warning if daemon_mode=false instead.
There could be existing misconfigurations for these in the wild that we
don't want to break by making working-but-incorrect configuration fail
with a fatal validation error instead.