Fixes to compile and run on Windows
Fix compilation on GOOS=windows for the main binary and tests,
both of which were broken.

Fix a leaked open file handle which caused config_test.go to fail
on windows with "Sharing violation" errors when cleaning up the
tempdir.

Add a 'test-wine' Makefile target. Explain how to run the tests
for windows via wine in the readme.

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>