Add linter fix hint and lint-fix make target
This project uses a private copy of golangci-lint stashed in .build,
so it can be confusing to have e.g. the linter fail with gofmt
errors, but a direct 'gofmt -w' invocation not fix them or 'gofmt -d'
not show them.

Direct the user to use the private copy of golangci-lint to run in --fix
mode and add a 'make lint-fix' target to help.

It would probably be better to use pre-commit for this, rather than a
make target, but this is less intrusive to existing workflows.

Example:

        $ make lint
	[snip]/pkg/sidecar/sidecar.go:35:6: var-naming: don't use underscores in Go names; func make_the_linter_sad should be makeTheLinterSad (revive)
	func make_the_linter_sad() {}
	     ^
	golangci-lint failed with 1; try make lint-fix or use [snip]/.build/linux-x86_64/golangci_lint/v1.64.4/golangci-lint to investigate
	make: *** [Makefile:174: lint-code] Error 1

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>