Fixes to compile and run on Windows (#257)
* Fixes to compile and run on Windows

Fix compilation on GOOS=windows for the main binary and tests,
both of which were broken.

Fix a leaked open file handle which caused config_test.go to fail
on windows with "Sharing violation" errors when cleaning up the
tempdir.

Add a 'test-wine' Makefile target. Explain how to run the tests
for windows via wine in the readme.

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>

* Update .gitignore

Newline added

Co-authored-by: Faisal Memon <fymemon@yahoo.com>
Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>

* Don't run config tests relating to pid on windows

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>

* Use go test -exec to run tests under wine

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>

* Remove extra Makefile whitespace

Signed-off-by: Faisal Memon <fymemon@yahoo.com>

---------

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>
Signed-off-by: Faisal Memon <fymemon@yahoo.com>
Co-authored-by: Faisal Memon <fymemon@yahoo.com>