Debug log on sidecar creation
This log message is more to ensure that the Log field is actually
defined in the config before we rely on it, but it's also useful
to understand what's happening in a test run.

Added because TestSignalProcess was creating a Config object without a
Log defined, so it passed unless the test run emitted a log message. In
that case it would crash with a surprising segfault.