Fix stdio handling in sidecar tests
The stdio redirection logic I added in the sidecar tests was incorrect,
and would always send output to os.DevNull. Fix so that tests default to
passthrough stdio like the real sidecar does. Add support for capturing
and logging stdout and stderr in cases where it should not be passed
through during test runs, e.g. when a command emits a "normal error" we
don't want the user to have to see.

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>