Add a test suite for sidecar cmd and signals (#258)
* Add a test suite for sidecar cmd and signals

Add test cover for:

* long-running and short-lived cmd and cmd_args invocations
* signalling via pid_file_name and renew_signal
* windows cmd and cmd_args execution

If rebased onto the 'fix-windows' branch for windows test fixes, these
pass for windows test runs too.

To make these tests possible the following changes are made to the
sidecar code:

* Sidecar.certReadyChan yields *workloadapi.X509Context not struct{}
  so the context can be observed. This could also be useful for
  code extending spiffe-helper.
* A new Sidecar.cmdExitChan channel reports whenever a process
  managed by spiffe-helper exits, and the exit state. This allows
  tests to wait for process exits and see the exit code, etc.
* A new Sidecar.pidFileSignalledChan channel reports whenever
  signalling the file in pid_file_name is attempted, the pid
  signalled, and the outcome. This allows signal based tests
  to see what spiffe-helper tried to signal.

Channels with values are used to ensure that there's no race between
what might be present in the Sidecar struct state when the test observes
it and what was there when the event of interest occurred.

It might be handy to have a channel for observing when the managed
process is started too, but the tests just busy-wait polling
Sidecar.processRunning to detect this for now.

I considered combining these channels into some kind of unified observer
channel that emits a union struct depending on the event type, but it
seemed pointless to do this when we can just select{} on multiple
channels of interest and keep each channel simple.

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>

* Add linter fix hint and lint-fix make target

This project uses a private copy of golangci-lint stashed in .build,
so it can be confusing to have e.g. the linter fail with gofmt
errors, but a direct 'gofmt -w' invocation not fix them or 'gofmt -d'
not show them.

Direct the user to use the private copy of golangci-lint to run in --fix
mode and add a 'make lint-fix' target to help.

It would probably be better to use pre-commit for this, rather than a
make target, but this is less intrusive to existing workflows.

Example:

        $ make lint
	[snip]/pkg/sidecar/sidecar.go:35:6: var-naming: don't use underscores in Go names; func make_the_linter_sad should be makeTheLinterSad (revive)
	func make_the_linter_sad() {}
	     ^
	golangci-lint failed with 1; try make lint-fix or use [snip]/.build/linux-x86_64/golangci_lint/v1.64.4/golangci-lint to investigate
	make: *** [Makefile:174: lint-code] Error 1

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>

* Skip TestSignalProcess on Windows

The current test implementation won't work on Windows. Someone's welcome
to adapt it, but for now this stops it from failing the test.

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>

* Debug log on sidecar creation

This log message is more to ensure that the Log field is actually
defined in the config before we rely on it, but it's also useful
to understand what's happening in a test run.

Added because TestSignalProcess was creating a Config object without a
Log defined, so it passed unless the test run emitted a log message. In
that case it would crash with a surprising segfault.

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>

* Fix stdio handling in sidecar tests

The stdio redirection logic I added in the sidecar tests was incorrect,
and would always send output to os.DevNull. Fix so that tests default to
passthrough stdio like the real sidecar does. Add support for capturing
and logging stdout and stderr in cases where it should not be passed
through during test runs, e.g. when a command emits a "normal error" we
don't want the user to have to see.

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>

* Fix long-running sidecar signal handling tests

New test TestSidecar_TestCmdRunsLongRunning didn't actually
assert that there was no signal received on the first launch,
so fix that.

Also clarify the comments to make the intent of the logic clearer, per
PR review.

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>

* Remove redundant test skip for Windows

The additions to TestSidecar_TestCmdRuns tested for and skipped windows
both at the start of the test and each iteration, which was redundant.

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>

* Clarify test comments about command specific timeouts.

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>

* Implement test hooks

Signed-off-by: Faisal Memon <fymemon@yahoo.com>

* Remove duplicated config struct

Signed-off-by: Faisal Memon <fymemon@yahoo.com>

* cleanup

Signed-off-by: Faisal Memon <fymemon@yahoo.com>

* Update pkg/sidecar/sidecar.go

Co-authored-by: Faisal Memon <fymemon@yahoo.com>
Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>

* Fix merge errors

Signed-off-by: Faisal Memon <fymemon@yahoo.com>

* Cleanup

Signed-off-by: Faisal Memon <fymemon@yahoo.com>

* Remove some debug logs

Signed-off-by: Faisal Memon <fymemon@yahoo.com>

* More cleanup

Signed-off-by: Faisal Memon <fymemon@yahoo.com>

* Fix error with process running at end of test

Signed-off-by: Faisal Memon <fymemon@yahoo.com>

* Fix merge error

Signed-off-by: Faisal Memon <fymemon@yahoo.com>

* Cleanup

* Update comment, remove loop assignments

Signed-off-by: Faisal Memon <fymemon@yahoo.com>

* Make if statement more readable

Signed-off-by: Faisal Memon <fymemon@yahoo.com>

---------

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>
Signed-off-by: Faisal Memon <fymemon@yahoo.com>
Co-authored-by: Faisal Memon <fymemon@yahoo.com>